Enhancement: ImportJobTest now uses an embedded MongoDb for all it's tests.
In other words, when this test is run, it now automatically creates an embedded MongoDb instance running on an unused port.It populates the embedded database with test data (read from json files in the test_data directory), runs all the ImportJob tests, then shuts down the embedded database. This change means you no longer need to have an instance of mongodb running, with all the right collections populated with test data, just to ensure that all the tests in dcc-release pass.